# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.0.1",
    "title": "Winding tree API",
    "description": "Node js API for Winding Tree",
    "contact": {
      "name": "Winding tree",
      "url": "https://windingtree.com/"
    },
    "license": {
      "name": "Apache-2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    }
  },
...
  "schemes": [
    "http"
  ],
  "host": "localhost:3000",
  "basePath": "/",
...
  "paths": {
...
    "/hotels/{hotelAddress}/unitTypes/{type}/units": {
      "post": {
        "tags": [ "Units" ],
        "summary": "Create units.",
        "description": "Add units to unit type.",
        "parameters": [
          {
            "in": "path",
            "name": "address",
            "type": "string",
            "required": true,
            "description": "Hotel address."
          },
          {
            "in": "path",
            "name": "type",
            "type": "string",
            "required": true,
            "description": "unit type."
          },
          {
            "in": "body",
            "name": "password",
            "required": true,
            "description": "Password.",
            "schema": {
              "required": [
                "password"
              ],
              "properties": {
                "password": {
                  "description": "Current password.",
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucefully created.",
            "schema": {
              "$ref": "#/definitions/txHash"
            }
          },
          "400": {
            "description": "Invalid input",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/hotels/{hotelAddress}/unitTypes/{type}/units/{unitAddress}" : {
      "delete": {
        "tags": [ "Units" ],
        "summary": "remove unit.",
        "description": "Remove unit for hotel.",
        "parameters": [
          {
            "in": "path",
            "name": "address",
            "type": "string",
            "required": true,
            "description": "Hotel address."
          },
          {
            "in": "path",
            "name": "type",
            "type": "string",
            "required": true,
            "description": "unit type."
          },
          {
            "in": "path",
            "name": "unit",
            "type": "string",
            "required": true,
            "description": "unit address."
          },
          {
            "in": "body",
            "name": "password",
            "required": true,
            "description": "Password.",
            "schema": {
              "required": [
                "password"
              ],
              "properties": {
                "password": {
                  "description": "Current password.",
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Sucefully removed.",
            "schema": {
              "$ref": "#/definitions/txHash"
            }
          },
          "400": {
            "description": "Invalid input",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/hotels/{hotelAddress}/unitTypes/{type}/units/{unitAddress}/active" : {
      "put": {
        "tags": [ "Units" ],
        "summary": "Active status.",
        "description": "Set active/inactive status for a unit of hotel.",
        "parameters": [
          {
            "in": "path",
            "name": "address",
            "type": "string",
            "required": true,
            "description": "Hotel address."
          },
          {
            "in": "path",
            "name": "type",
            "type": "string",
            "required": true,
            "description": "unit type."
          },
          {
            "in": "path",
            "name": "unit",
            "type": "string",
            "required": true,
            "description": "unit address."
          },
          {
            "in": "body",
            "name": "params",
            "required": true,
            "description": "parameters.",
            "schema": {
              "required": [
                "password",
                "active"
              ],
              "properties": {
                "password": {
                  "description": "Current password.",
                  "type": "string"
                },
                "active": {
                  "description": "Sets unit active/inactive.",
                  "type": "boolean"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucefully updated.",
            "schema": {
              "$ref": "#/definitions/txHash"
            }
          },
          "400": {
            "description": "Invalid input",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/hotels/{hotelAddress}/units/{unitAddress}/book" : {
      "put": {
        "tags": [ "Units" ],
        "summary": "Book an unit.",
        "description": "Book a unit.",
        "parameters": [
          {
            "in": "path",
            "name": "hotelAddress",
            "type": "string",
            "required": true,
            "description": "Hotel address."
          },
          {
            "in": "path",
            "name": "unitAddress",
            "type": "string",
            "required": true,
            "description": "unit address."
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "description": "parameters.",
            "schema": {
              "required": [
                "account",
                "guest",
                "days",
                "from"
              ],
              "properties": {
                "account": {
                  "description": "Client's account address.",
                  "type": "string"
                },
                "guest": {
                  "description": "Guest data.",
                  "type": "string"
                },
                "days": {
                  "description": "Amount of days.",
                  "type": "number"
                },
                "from": {
                  "description": "First day of the booking.",
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucefully updated.",
            "schema": {
              "$ref": "#/definitions/txHash"
            }
          },
          "400": {
            "description": "Invalid input",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
     "/units/{unitAddress}/available":{
         "get":{
           "tags": [ "Units" ],
         "summary": "Unit availability.",
           "description": "Get unit availability for a range of days.",
           "parameters": [
             {
               "in": "path",
               "name": "unit",
               "type": "string",
               "required": true,
               "description": "Unit address."
             },
             {
               "name": "params",
               "in": "body",
               "schema": {
                 "required": [
                   "from",
                   "days"
                 ],
                 "properties": {
                   "from": {
                     "description": "Start date. UTC day since 1970",
                     "type": "number"
                   },
                   "days": {
                     "description": "Amount of days of the reservation.",
                     "type": "numer"
                   }
                 }
               }
             }
           ],
           "responses": {
             "200": {
               "description": "Ok.",
               "schema": {
                 "properties": {
                   "available": {
                     "description": "True if the unit is available.",
                     "type": "boolean"
                   }
                 }
               }
             },
             "400": {
               "description": "Can not get.",
               "schema": {
                 "$ref": "#/definitions/Error"
               }
             }
           }
         }
       },
...
  },
  "definitions": {
    "txHash": {
      "title": "Transaction hash",
      "description": "transaction hash for the operation.",
      "properties": {
        "txHash": {
          "type": "string",
          "description": "Transaction hash."
        }
      }
    },
    "Error": {
      "title": "Error",
      "description": "Default schema for errors returned by API.",
      "properties": {
        "code": {
          "type": "string",
          "description": "Computer-readable error. Use this for comparison."
        },
        "short": {
          "type": "string",
          "description": "Human-readable error with a short description of the error."
        },
        "long": {
          "type": "string",
          "description": "Verbose explanation of what happen."
        }
      }
    },
...
  }
}